#9 refactor: ApiResponse 파일을 response 패키지에 넣음, AccommodationController에도 ApiResponse 적용 완료
