--- src/main/java/com/backoffice/upjuyanolja/global/common/response/Status.java
@@ -1,5 +1,0 @@
-package com.backoffice.upjuyanolja.global.common.response;
-
-public enum Status {
-    SUCCESS, FAIL, ERROR
-}
